Made semantic version components unsigned

+++ src/version.c
@@ -1,11 +1,12 @@
 #include "libgitledger/version.h"
 
 #include <stddef.h>
+#include <stdint.h>
 
 static const gitledger_semantic_version_t GITLEDGER_VERSION_VALUE = {0, 1, 0};
 static char                               gitledger_version_buffer[16];
 
-static int write_decimal(unsigned int value, char** cursor, size_t* remaining)
+static int write_decimal(uint32_t value, char** cursor, size_t* remaining)
 {
     char   digits[10];
     size_t idx = 0U;
@@ -46,7 +47,7 @@ const char* gitledger_semantic_version_string(void)
     char*  cursor    = gitledger_version_buffer;
     size_t remaining = sizeof(gitledger_version_buffer);
 
-    if (!write_decimal((unsigned int) GITLEDGER_VERSION_VALUE.major, &cursor, &remaining))
+    if (!write_decimal(GITLEDGER_VERSION_VALUE.major, &cursor, &remaining))
         {
             return "";
         }
@@ -59,7 +60,7 @@ const char* gitledger_semantic_version_string(void)
     *cursor++ = '.';
     remaining--;
 
-    if (!write_decimal((unsigned int) GITLEDGER_VERSION_VALUE.minor, &cursor, &remaining))
+    if (!write_decimal(GITLEDGER_VERSION_VALUE.minor, &cursor, &remaining))
         {
             return "";
         }
@@ -72,7 +73,7 @@ const char* gitledger_semantic_version_string(void)
     *cursor++ = '.';
     remaining--;
 
-    if (!write_decimal((unsigned int) GITLEDGER_VERSION_VALUE.patch, &cursor, &remaining))
+    if (!write_decimal(GITLEDGER_VERSION_VALUE.patch, &cursor, &remaining))
         {
             return "";
         }
